refactor(web): modularize API key dialogs and form components
